# csharp-cosmos/src/api/wwwroot/openapi.yaml
openapi: 3.0.0
info:
  description: Minimal scaffold API
  version: 1.0.0
  title: Backend API
  contact:
    email: azdevteam@microsoft.com

paths:
  /:
    get:
      operationId: GetRoot
      summary: Root endpoint
      responses:
        200:
          description: OK
  /health:
    get:
      operationId: GetHealth
      summary: Health check
      responses:
        200:
          description: OK
  /api/v1/ping:
    get:
      operationId: GetPingV1
      summary: Authenticated ping probe
      security:
        - bearerAuth: []
      responses:
        200:
          description: OK
        401:
          description: Missing or invalid JWT
        403:
          description: Missing required permission
  /api/v1/rbac-probe:
    get:
      operationId: GetRbacProbe
      summary: Requires licenses:write permission
      security:
        - bearerAuth: []
      responses:
        200:
          description: OK
        401:
          description: Missing or invalid JWT
        403:
          description: Missing required permission
  /api/v1/auth/login:
    post:
      operationId: PostAuthLogin
      summary: Login (bootstrap users)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
      responses:
        200:
          description: Access and refresh tokens
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      accessToken:
                        type: string
                      refreshToken:
                        type: string
                      expiresInSeconds:
                        type: integer
        401:
          description: Invalid credentials
  /api/v1/auth/refresh:
    post:
      operationId: PostAuthRefresh
      summary: Rotate refresh token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refreshToken]
              properties:
                refreshToken:
                  type: string
      responses:
        200:
          description: New token pair
        401:
          description: Invalid or expired refresh token
  /api/v1/auth/logout:
    post:
      operationId: PostAuthLogout
      summary: Revoke refresh token (requires Bearer access token)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refreshToken]
              properties:
                refreshToken:
                  type: string
      responses:
        204:
          description: Logged out
        401:
          description: Not authenticated
        403:
          description: Refresh token does not match current user

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT